ai_verifier/requirements.txt
langchain>=0.1.0
langchain-groq>=0.1.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
